Wave System Refactored
Turned the entire wave system into a state machine, a lot of it is just temporary until I get hold of the actual assets we're gonna be using.

Part of this was creating a base for state machines to use, if anyone ever plans on making a state machine, the base is now here and ready to be used, just ask and I'll help you get things set up.